Remove lib/ source prefix
This prefix was a holdover from early days when Python
standard practices were not as clear and we had come from
a Perl background as far as scripting languages were concerned.

Specifically, this document:

https://docs.python.org/2/distutils/setupscript.html#listing-whole-packages

suggested using lib/ was perhaps still a common thing, and maybe it was.
But soon after, it was definitely not.  git seems to travel well as far
as things like cherry-picks and such (also not a thing we had back then).

Change-Id: Id528da1c8ec902b5d50fb360af62db366b03b419